+++ proto/storage/config.proto
@@ -142,11 +142,11 @@ message PrometheusMetricsConfig {
     map<string, LabelExpressions> metric_labels = 2;
   }
 
-  message PolicyViolations {
+  message Alerts {
     uint32 gathering_period_hours = 1;
     map<string, LabelExpressions> metric_labels = 2;
   }
 
   Vulnerabilities vulnerabilities = 1;
-  PolicyViolations policy_violations = 2;
+  Alerts alerts = 2;
 }
